+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Dialog & Lightbox: fixed focus trap not working when closing and reopening the same component.
 -   Dialog & Lightbox & DatePickerField: fixed focus trap on nested components.
 -   Dialog & Lightbox & Popover: fixed 'Escape' key to close nested components.
+-   DatePickerField: fix focus element when opened and re-opened.
 
 ## [2.2.19][] - 2022-06-03